@@ -9,117 +9,130 @@
     "build": "gatsby build",
     "develop": "gatsby develop",
     "format": "prettier --write 'src/**/*.js'",
-    "lint": "eslint 'src/**/*.js'",
-    "lint-fix": "eslint --fix 'src/**/*.js'",
+    "lint": "eslint 'src/**/*.[jt]s?(x)' && remark . --frail",
+    "lint-fix": "eslint --fix 'src/**/*.[jt]s?(x)'",
     "test": "yarn jest --runInBand && codecov",
-    "test:watch": "yarn jest --watch"
-  },
-  "jest": {
-    "moduleNameMapper": {
-      "\\.(css|jpg|png|svg)$": "<rootDir>/empty-module.js"
-    },
-    "testMatch": [
-      "**/?(*_)(spec|test).js"
-    ],
-    "setupFiles": [
-      "./src/configuration/jest.config.js"
-    ],
-    "snapshotSerializers": [
-      "enzyme-to-json/serializer"
-    ],
-    "coverageDirectory": "./coverage/",
-    "collectCoverage": true,
-    "globals": {
-      "__PATH_PREFIX__": ""
-    }
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged "
-    }
+    "test:watch": "yarn jest --watch",
+    "precommit": "lint-staged",
+    "storybook": "start-storybook -p 9001 -c storybook"
   },
   "lint-staged": {
-    "src/**/*.{js,json}": [
-      "eslint"
-    ]
-  },
-  "resolutions": {
-    "babel-core": "7.0.0-bridge.0"
+    "src/**/*.{js,json,ts,tsx}": [
+      "eslint --fix --plugin tsc --rule 'tsc/config: [2, {configFile: \"tsconfig.json\"}]'",
+      "stylelint",
+      "jest --findRelatedTests",
+      "git add"
+    ],
+    "*.md": "remark . --frail"
   },
   "dependencies": {
+    "@babel/preset-typescript": "^7.7.4",
     "@rebass/grid": "^6.0.0",
+    "@types/storybook__addon-info": "^5.2.1",
     "debounce": "^1.2.0",
-    "dotenv": "^8.0.0",
-    "eslint": "^5.16.0",
-    "gatsby": "^2.13.59",
-    "gatsby-image": "^2.2.8",
-    "gatsby-link": "^2.2.3",
-    "gatsby-plugin-layout": "^1.1.3",
-    "gatsby-plugin-offline": "^2.2.5",
-    "gatsby-plugin-react-helmet": "^3.1.3",
+    "dotenv": "^8.2.0",
+    "gatsby": "^2.19.1",
+    "gatsby-image": "^2.2.39",
+    "gatsby-link": "^2.2.28",
+    "gatsby-plugin-layout": "^1.1.21",
+    "gatsby-plugin-offline": "^3.0.32",
+    "gatsby-plugin-react-helmet": "^3.1.21",
     "gatsby-plugin-robots-txt": "^1.5.0",
     "gatsby-plugin-rollbar": "^1.0.0",
-    "gatsby-plugin-sharp": "^2.2.11",
-    "gatsby-plugin-sitemap": "^2.2.6",
-    "gatsby-plugin-styled-components": "^3.1.2",
-    "gatsby-source-contentful": "^2.1.19",
-    "gatsby-source-filesystem": "^2.1.9",
-    "gatsby-transformer-sharp": "^2.2.6",
+    "gatsby-plugin-sharp": "^2.4.0",
+    "gatsby-plugin-sitemap": "^2.2.26",
+    "gatsby-plugin-styled-components": "^3.1.18",
+    "gatsby-source-contentful": "^2.1.78",
+    "gatsby-source-filesystem": "^2.1.45",
+    "gatsby-transformer-sharp": "^2.3.13",
     "js-cookie": "^2.2.1",
     "lodash.debounce": "^4.0.8",
     "moment": "^2.23.0",
     "no-scroll": "^2.1.1",
-    "polished": "^3.4.0",
+    "polished": "^3.4.2",
     "querystring": "^0.2.0",
-    "react-accessible-accordion": "^2.4.5",
+    "react-accessible-accordion": "^3.0.1",
     "react-dates": "16.7.0",
-    "react-flip-move": "^3.0.3",
     "react-helmet": "^5.2.1",
     "react-image-gallery": "^0.8.18",
     "react-intercom": "^1.0.15",
-    "react-markdown": "^4.1.0",
-    "react-onclickoutside": "^6.8.0",
-    "react-slick": "^0.24.0",
-    "react-test-renderer": "^16.8.6",
+    "react-markdown": "^4.2.2",
+    "react-onclickoutside": "^6.9.0",
+    "react-slick": "^0.25.2",
+    "react-test-renderer": "^16.10.1",
     "react-typeform-embed": "^0.2.1",
-    "shortid": "^2.2.14",
+    "shortid": "^2.2.15",
     "slick-carousel": "^1.8.1",
+    "slugify": "^1.3.6",
     "string-hash": "^1.1.3",
-    "styled-components": "^4.3.2",
-    "tinycolor2": "^1.4.1"
+    "styled-components": "^4.4.1",
+    "styled-system": "^5.1.4",
+    "tinycolor2": "^1.4.1",
+    "uuid-encoder": "^1.2.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.5.5",
+    "@babel/core": "^7.6.0",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/preset-env": "^7.5.5",
+    "@babel/preset-env": "^7.6.0",
     "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.7.4",
     "@contentful/rich-text-react-renderer": "^13.2.0",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "^10.0.1",
-    "babel-plugin-remove-graphql-queries": "^2.6.3",
+    "@storybook/addon-a11y": "~5.2.8",
+    "@storybook/addon-actions": "~5.2.8",
+    "@storybook/addon-info": "~5.1.11",
+    "@storybook/addon-knobs": "~5.2.8",
+    "@storybook/react": "~5.2.8",
+    "@types/jest": "^24.0.22",
+    "@types/react": "^16.9.14",
+    "@types/react-dom": "^16.9.4",
+    "@types/styled-components": "^4.4.0",
+    "@types/styled-system": "^5.1.4",
+    "@typescript-eslint/eslint-plugin": "^2.8.0",
+    "@typescript-eslint/parser": "^2.8.0",
+    "babel-eslint": "^10.0.3",
+    "babel-loader": "^8.0.6",
+    "babel-plugin-remove-graphql-queries": "^2.7.10",
     "babel-plugin-styled-components": "^1.10.6",
     "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-preset-react-app": "^9.1.0",
     "codecov": "^3.2.0",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.14.0",
     "enzyme-to-json": "^3.4.0",
-    "eslint": "^5.16.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-config-prettier": "^6.0.0",
+    "eslint": "^6.7.1",
+    "eslint-config-airbnb": "^18.0.1",
+    "eslint-config-prettier": "^6.7.0",
+    "eslint-config-react-app": "^5.0.2",
+    "eslint-plugin-flowtype": "^4.3.0",
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.0.1",
+    "eslint-plugin-prettier": "^3.1.1",
     "eslint-plugin-react": "^7.14.3",
-    "eslint-plugin-react-hooks": "^1.6.1",
-    "husky": "^3.0.3",
-    "jest": "^23.6.0",
-    "jest-environment-jsdom": "^24.8.0",
+    "eslint-plugin-react-hooks": "^2.3.0",
+    "eslint-plugin-tsc": "^1.2.0",
+    "gatsby-plugin-typescript": "^2.1.15",
+    "husky": "^3.1.0",
+    "jest": "^24.9.0",
+    "jest-environment-jsdom": "^24.0.0",
     "jest-styled-components": "^6.3.1",
     "jest-transform-stub": "^2.0.0",
-    "lint-staged": "^8.2.1",
+    "lint-staged": "^9.4.3",
     "prettier": "^1.18.2",
     "prop-types": "^15.6.2",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6"
+    "react": "^16.9.0",
+    "react-docgen-typescript-loader": "^3.6.0",
+    "react-dom": "^16.12.0",
+    "remark-cli": "^7.0.1",
+    "remark-preset-lint-recommended": "^3.0.3",
+    "stylelint": "^12.0.0",
+    "stylelint-a11y": "^1.2.2",
+    "stylelint-config-prettier": "^6.0.0",
+    "stylelint-config-recommended": "^3.0.0",
+    "stylelint-config-styled-components": "^0.1.1",
+    "stylelint-prettier": "^1.1.1",
+    "stylelint-processor-styled-components": "^1.9.0",
+    "ts-loader": "^6.2.1",
+    "typescript": "^3.7.2",
+    "typescript-styled-plugin": "^0.15.0"
   }
 }
