@@ -1,12 +1,12 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { Container, Row } from '../grid'
+import Button from '../../components/button'
 import {
   CTAWrapper,
   CTAPanel,
   CTATitle,
   CTABody,
-  CTAButton,
   RelativeColumn,
 } from './styles'
 
@@ -20,11 +20,7 @@ const CTABox = props => {
             <CTAPanel>
               {title && <CTATitle>{title}</CTATitle>}
               {body && <CTABody>{body}</CTABody>}
-              {link.to && (
-                <CTAButton to={link.to} primary>
-                  {link.text}
-                </CTAButton>
-              )}
+              {link.to && <Button to={link.to}>{link.text}</Button>}
             </CTAPanel>
           </RelativeColumn>
         </Row>
