@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 import PropTypes from 'prop-types'
-import { filterTypes } from '../../constants'
+import constants from '../../constants'
 import FilterContainer, { calculateIsSelected } from '../filterContainer'
 import PagedCardContainer from '../pagedCardContainer'
 
@@ -10,6 +10,8 @@ export const calculateInitialSelected = (filterType, showAllCategoryTitle) => {
       return showAllCategoryTitle ? [showAllCategoryTitle] : []
     case 'radio':
       return showAllCategoryTitle
+    default:
+      return
   }
 }
 
@@ -42,6 +44,8 @@ export const calculateSelected = (
     }
     case 'radio':
       return filterNameSelected
+    default:
+      return
   }
 }
 
@@ -61,6 +65,8 @@ export const calculateShouldShowCard = (
       return selected === showAllCategoryTitle
         ? true
         : category.includes(selected)
+    default:
+      return
   }
 }
 
@@ -86,7 +92,7 @@ export const calculateAvailableCategories = (
     : availableCategories
 }
 
-const FilteredCardContainer = ({
+const FilteredPagedCardContainer = ({
   filterType,
   categories,
   showAllCategoryTitle,
@@ -150,8 +156,8 @@ const FilteredCardContainer = ({
   )
 }
 
-FilteredCardContainer.propTypes = {
-  filterType: PropTypes.oneOf(filterTypes).isRequired,
+FilteredPagedCardContainer.propTypes = {
+  filterType: PropTypes.oneOf(constants.filterTypes).isRequired,
   categories: PropTypes.arrayOf(
     PropTypes.shape({ title: PropTypes.string, hexColour: PropTypes.string })
   ).isRequired,
@@ -164,10 +170,10 @@ FilteredCardContainer.propTypes = {
   showMoreButtonText: PropTypes.string,
 }
 
-FilteredCardContainer.defaultProps = {
+FilteredPagedCardContainer.defaultProps = {
   pageSize: 6,
   showAllCategoryTitle: '',
   showMoreButtonText: 'Show more',
 }
 
-export default FilteredCardContainer
+export default FilteredPagedCardContainer
