@@ -10,6 +10,7 @@ import iconClose from '../../../theme/assets/images/icon-close.svg'
 import EventDateFilter from '../filters/eventDateFilter'
 import EventFreeFilter from '../filters/eventFreeFilter'
 import EventDropdownFilter from '../filters/eventDropdownFilter'
+import Button from '../../../components/button'
 
 const FilterWrapper = styled(Flex)`
   background-color: ${theme.colors.white};
@@ -146,9 +147,11 @@ const spin = keyframes`
 const animation = props => {
   if (props.clickAnimation)
     return css`
-      ${spin} 0.3s ease-in-out
+      animation: ${spin} 0.3s ease-in-out;
     `
-  return css`unset`
+  return css`
+    animation: unset;
+  `
 }
 /* eslint-enable */
 
@@ -161,7 +164,7 @@ const IconClear = styled.i`
   position: absolute;
   left: 0;
   top: 2px;
-  animation: ${animation};
+  ${animation}
 
   &:before {
     content: '';
@@ -191,6 +194,19 @@ const CloseButton = styled.button`
     display: none;
   `};
 `
+
+const SubmitButtonContainer = styled.div`
+  margin: 20px 0 -20px 0;
+  width: 100%;
+  padding: 20px;
+  box-shadow: 0 -3px 5px 0 rgba(0, 0, 0, 0.1);
+  height: 90px;
+
+  ${media.tablet`
+    display: none;
+  `};
+`
+
 class EventsFilters extends Component {
   state = {
     clickAnimation: false,
@@ -299,6 +315,15 @@ class EventsFilters extends Component {
             <FlexColumn width={[1, 1, 0.5, 0.3333]}>
               <EventFreeFilter />
             </FlexColumn>
+            <SubmitButtonContainer>
+              <Button
+                width={{ default: '100%', md: 'auto' }}
+                onClick={this.props.toggleFiltersMobile}
+              >
+                Show {context.filteredEvents.length} event
+                {context.filteredEvents.length !== 1 && 's'}
+              </Button>
+            </SubmitButtonContainer>
           </FilterWrapper>
         )}
       </Consumer>
