@@ -1,9 +1,8 @@
 export { default as EventListingCard } from './components/eventListingCard'
+export { default as GroupedEventsCards } from './components/groupedEventsCard'
 export { default as EventTagList } from './components/eventTagList'
 export { default as EventSchedule } from './components/eventSchedule'
 export { default as EventsYouMayLike } from './components/eventsYouMayLike'
 export { default as EventInfoCard } from './components/eventInfoCard'
-export {
-  default as EventDirectionsSection,
-} from './components/eventDirectionsSection'
+export { default as EventDirectionsSection } from './components/eventDirectionsSection'
 export { default as EventShareSection } from './components/eventShareSection'
