@@ -2,7 +2,7 @@ import { StaticQuery, graphql } from 'gatsby'
 import React from 'react'
 import moment from 'moment'
 import { getDuration, filterPastEvents } from '../../../events/helpers'
-import { dateFormat } from '../../../../constants'
+import constants from '../../../../constants'
 import FeaturedEvents from '.'
 
 const query = graphql`
@@ -46,7 +46,7 @@ const FeaturedEventsQuery = props => (
 
           const recurrenceDates = eventRecurrenceDates
             .map(date => {
-              const startTime = moment(date, dateFormat)
+              const startTime = moment(date, constants.dateFormat)
                 .hours(originalStartTime.hours())
                 .minutes(originalStartTime.minutes())
                 .toISOString()
@@ -69,7 +69,7 @@ const FeaturedEventsQuery = props => (
           const customId =
             recurrenceDates.length > 0
               ? `${event.node.id}-${moment(startTime)
-                  .format(dateFormat)
+                  .format(constants.dateFormat)
                   .split('/')
                   .join('')}`
               : event.node.id
@@ -90,4 +90,12 @@ const FeaturedEventsQuery = props => (
   />
 )
 
+/**
+ * @deprecated
+ * This needs updated following Jira ticket EVE-2.
+ * The way we handle event IDs has changed, specifically with recurring events.
+ * We no longer rewrite the ID on a per occurrence basis, so there are multiple events existing with the same ID.
+ * This will cause issues with the customId generation.
+ * Point of contact: Em McDonald
+ */
 export default FeaturedEventsQuery
