@@ -3,6 +3,7 @@ import PropTypes from 'prop-types'
 import { StaticQuery, graphql } from 'gatsby'
 import Helmet from 'react-helmet'
 import styled, { ThemeProvider } from 'styled-components'
+import HelmetComponet from '../components/helmetComponent'
 import { Provider } from '../components/appContext'
 import CookieNotice from '../components/cookieNotice'
 import Nav from '../components/navigation/nav'
@@ -29,9 +30,8 @@ import logo from '../theme/assets/images/logo-pride.svg'
 import theme from '../theme/theme'
 import 'slick-carousel/slick/slick.css'
 import 'slick-carousel/slick/slick-theme.css'
-import './index.css'
-import './fonts.css'
 import metaImg from '../theme/assets/images/pride-logo-social.png'
+import { GlobalStyle } from './index.styles'
 
 const SiteWrapper = styled.div`
   max-width: 1600px;
@@ -98,6 +98,7 @@ const query = graphql`
 
 const Layout = ({ children, location: { pathname } }) => (
   <ThemeProvider theme={theme}>
+    <GlobalStyle />
     <StaticQuery
       query={query}
       render={({
@@ -130,14 +131,9 @@ const Layout = ({ children, location: { pathname } }) => (
         return (
           <Provider events={events}>
             <Fragment>
+              <HelmetComponet title={title} />
               <Helmet
-                title={title}
                 meta={[
-                  {
-                    name: 'description',
-                    content: description,
-                  },
-
                   // Schema meta tags for http://schema.org/WebPage
                   {
                     itemprop: 'name',
