@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React from 'react'
 import styled from 'styled-components'
 import PropTypes from 'prop-types'
 import { graphql } from 'gatsby'
@@ -99,264 +99,256 @@ const HeroImage = styled(Img)`
   width: 100% !important;
 `
 
-// eslint-disable-next-line react/prefer-stateless-function
-export default class Event extends Component {
-  render() {
-    const {
-      data: {
-        contentfulEvent: {
-          id,
-          individualEventPicture,
-          eventDescription: { eventDescription },
-          name,
-          performances,
-          eventPriceLow,
-          eventCategories,
-          ticketingUrl,
-          accessibilityDetails,
-          location,
-          locationName,
-          addressLine1,
-          addressLine2,
-          city,
-          postcode,
-          eventSponsorSection,
-        },
-        contentfulEvent,
-        site: {
-          siteMetadata: { siteUrl },
-        },
-      },
-      location: { pathname },
-      pageContext: { startTime, endTime },
-      pageContext,
-    } = this.props
+const Event = ({
+  data: {
+    contentfulEvent: {
+      id,
+      individualEventPicture,
+      eventDescription: { eventDescription },
+      name,
+      performances,
+      eventPriceLow,
+      eventCategories,
+      ticketingUrl,
+      accessibilityDetails,
+      location,
+      locationName,
+      addressLine1,
+      addressLine2,
+      city,
+      postcode,
+      eventSponsorSection,
+    },
+    contentfulEvent,
+    site: {
+      siteMetadata: { siteUrl },
+    },
+  },
+  location: { pathname },
+  pageContext: { startTime, endTime },
+  pageContext,
+}) => {
+  const metaImg = `https:${individualEventPicture.file.url}?w=1000&h=562`
+  const metaUrl = siteUrl + pathname
 
-    const metaImg = `https:${individualEventPicture.file.url}?w=1000&h=562`
-    const metaUrl = siteUrl + pathname
+  const sponsorSection = eventSponsorSection
+    ? {
+        ...eventSponsorSection,
+        sponsors: eventSponsorSection.sponsors.map(sponsor => ({
+          name: sponsor.sponsorName,
+          url: sponsor.sponsorUrl,
+          logo: sponsor.sponsorLogo && sponsor.sponsorLogo.sizes.src,
+          level: sponsor.sponsorLevel,
+        })),
+      }
+    : null
 
-    const sponsorSection = eventSponsorSection
-      ? {
-          ...eventSponsorSection,
-          sponsors: eventSponsorSection.sponsors.map(sponsor => ({
-            name: sponsor.sponsorName,
-            url: sponsor.sponsorUrl,
-            logo: sponsor.sponsorLogo && sponsor.sponsorLogo.sizes.src,
-            level: sponsor.sponsorLevel,
-          })),
-        }
-      : null
+  return (
+    <PageWrapper>
+      <Helmet
+        title={name}
+        meta={[
+          {
+            name: 'description',
+            content: eventDescription,
+          },
 
-    return (
-      <PageWrapper>
-        <Helmet
-          title={name}
-          meta={[
-            {
-              name: 'description',
-              content: eventDescription,
-            },
+          // Schema meta tags http://schema.org/WebPage
+          {
+            itemprop: 'name',
+            content: name,
+          },
+          {
+            itemprop: 'description',
+            content: eventDescription,
+          },
+          {
+            itemprop: 'url',
+            content: metaUrl,
+          },
+          {
+            itemprop: 'thumbnailUrl',
+            content: metaImg,
+          },
+          {
+            itemprop: 'image',
+            content: metaImg,
+          },
 
-            // Schema meta tags http://schema.org/WebPage
-            {
-              itemprop: 'name',
-              content: name,
-            },
-            {
-              itemprop: 'description',
-              content: eventDescription,
-            },
-            {
-              itemprop: 'url',
-              content: metaUrl,
-            },
-            {
-              itemprop: 'thumbnailUrl',
-              content: metaImg,
-            },
-            {
-              itemprop: 'image',
-              content: metaImg,
-            },
+          // OpenGraph Meta Tags
+          {
+            property: 'og:title',
+            content: name,
+          },
+          {
+            property: 'og:description',
+            content: eventDescription,
+          },
+          {
+            property: 'og:latitude',
+            content: location.lat,
+          },
+          {
+            property: 'og:longitude',
+            content: location.lon,
+          },
+          {
+            property: 'og:street-address',
+            content: !addressLine1
+              ? ''
+              : addressLine2
+              ? `${locationName}, ${addressLine1}, ${addressLine2}`
+              : `${locationName}, ${addressLine1}`,
+          },
+          {
+            property: 'og:locality',
+            content: city,
+          },
+          {
+            property: 'og:postal-code',
+            content: postcode,
+          },
+          {
+            property: 'og:url',
+            content: metaUrl,
+          },
+          {
+            property: 'og:image',
+            content: metaImg,
+          },
+          {
+            property: 'og:image:secure_url',
+            content: metaImg,
+          },
 
-            // OpenGraph Meta Tags
-            {
-              property: 'og:title',
-              content: name,
-            },
-            {
-              property: 'og:description',
-              content: eventDescription,
-            },
-            {
-              property: 'og:latitude',
-              content: location.lat,
-            },
-            {
-              property: 'og:longitude',
-              content: location.lon,
-            },
-            {
-              property: 'og:street-address',
-              content: !addressLine1
-                ? ''
-                : addressLine2
-                ? `${locationName}, ${addressLine1}, ${addressLine2}`
-                : `${locationName}, ${addressLine1}`,
-            },
-            {
-              property: 'og:locality',
-              content: city,
-            },
-            {
-              property: 'og:postal-code',
-              content: postcode,
-            },
-            {
-              property: 'og:url',
-              content: metaUrl,
-            },
-            {
-              property: 'og:image',
-              content: metaImg,
-            },
-            {
-              property: 'og:image:secure_url',
-              content: metaImg,
-            },
-
-            // Twitter Meta Tags
-            {
-              name: 'twitter:title',
-              content: name,
-            },
-            {
-              name: 'twitter:description',
-              content: eventDescription,
-            },
-            {
-              name: 'twitter:image',
-              content: metaImg,
-            },
-            {
-              name: 'twitter:url',
-              content: metaUrl,
-            },
-            {
-              name: 'twitter:card',
-              content: 'summary_large_image',
-            },
-          ]}
-          link={[
-            {
-              rel: 'image_src',
-              content: metaImg,
-            },
-          ]}
-          script={[
-            {
-              type: 'application/ld+json',
-              innerHTML: JSON.stringify({
-                '@context': 'https://www.schema.org',
-                '@type': 'Event',
-                name,
-                url: metaUrl,
-                description: eventDescription,
-                startDate: startTime,
-                endDate: endTime,
-                image: metaImg,
-                location: {
-                  '@type': 'Place',
-                  name: "Joe's Party Palace",
-                  address: {
-                    '@type': 'PostalAddress',
-                    streetAddress: addressLine1,
-                    ...(addressLine2 && { addressLocality: addressLine2 }),
-                    addressRegion: city,
-                    postalCode: postcode,
-                    addressCountry: 'UK',
-                  },
+          // Twitter Meta Tags
+          {
+            name: 'twitter:title',
+            content: name,
+          },
+          {
+            name: 'twitter:description',
+            content: eventDescription,
+          },
+          {
+            name: 'twitter:image',
+            content: metaImg,
+          },
+          {
+            name: 'twitter:url',
+            content: metaUrl,
+          },
+          {
+            name: 'twitter:card',
+            content: 'summary_large_image',
+          },
+        ]}
+        link={[
+          {
+            rel: 'image_src',
+            content: metaImg,
+          },
+        ]}
+        script={[
+          {
+            type: 'application/ld+json',
+            innerHTML: JSON.stringify({
+              '@context': 'https://www.schema.org',
+              '@type': 'Event',
+              name,
+              url: metaUrl,
+              description: eventDescription,
+              startDate: startTime,
+              endDate: endTime,
+              image: metaImg,
+              location: {
+                '@type': 'Place',
+                name: "Joe's Party Palace",
+                address: {
+                  '@type': 'PostalAddress',
+                  streetAddress: addressLine1,
+                  ...(addressLine2 && { addressLocality: addressLine2 }),
+                  addressRegion: city,
+                  postalCode: postcode,
+                  addressCountry: 'UK',
                 },
-                offers: {
-                  '@type': 'Offer',
-                  price: eventPriceLow,
-                  priceCurrency: 'GBP',
-                  ...(ticketingUrl && { url: ticketingUrl }),
-                },
-              }),
-            },
-          ]}
+              },
+              offers: {
+                '@type': 'Offer',
+                price: eventPriceLow,
+                priceCurrency: 'GBP',
+                ...(ticketingUrl && { url: ticketingUrl }),
+              },
+            }),
+          },
+        ]}
+      />
+      <HeroImageAndTitle>
+        <HeroImage
+          aria-hidden="true"
+          objectFit="cover"
+          objectPosition="50% 50%"
+          fixed={
+            !checkBreakpoint(400)
+              ? individualEventPicture.mobile
+              : !checkBreakpoint(800)
+              ? individualEventPicture.tablet
+              : individualEventPicture.desktop
+          }
+          role="presentation"
         />
-        <HeroImageAndTitle>
-          <HeroImage
-            aria-hidden="true"
-            objectFit="cover"
-            objectPosition="50% 50%"
-            fixed={
-              !checkBreakpoint(400)
-                ? individualEventPicture.mobile
-                : !checkBreakpoint(800)
-                ? individualEventPicture.tablet
-                : individualEventPicture.desktop
-            }
-            role="presentation"
-          />
-        </HeroImageAndTitle>
-        <EventInfoCardWrapper>
-          <Container>
-            <Row>
-              <RelativeColumn width={1}>
-                <EventInfoCard
-                  data={contentfulEvent}
-                  pageContext={pageContext}
-                />
-              </RelativeColumn>
-            </Row>
-          </Container>
-        </EventInfoCardWrapper>
+      </HeroImageAndTitle>
+      <EventInfoCardWrapper>
         <Container>
           <Row>
-            <Column width={[1, 1, 0.6]}>
-              <TitleWrapper>
-                <Title>{name}</Title>
-                <EventTagList values={eventCategories} />
-              </TitleWrapper>
-              <Section>
-                <ReactMarkdown source={eventDescription} />
-              </Section>
-              {sponsorSection && (
-                <SponsorsSubsection title={sponsorSection.displayName}>
-                  {renderSponsors(sponsorSection.sponsors, true)}
-                </SponsorsSubsection>
-              )}
-              {accessibilityDetails && (
-                <>
-                  <AccessibilityHeading>Accessibility</AccessibilityHeading>
-                  <Section>
-                    <ReactMarkdown
-                      source={accessibilityDetails.accessibilityDetails}
-                    />
-                  </Section>
-                </>
-              )}
-              {performances && (
-                <Section>
-                  <EventSchedule schedule={performances} />
-                </Section>
-              )}
-              <EventShareSection
-                name={name}
-                location={metaUrl}
-                date={startTime}
-              />
-            </Column>
+            <RelativeColumn width={1}>
+              <EventInfoCard data={contentfulEvent} pageContext={pageContext} />
+            </RelativeColumn>
           </Row>
         </Container>
-        <EventDirectionsSection data={contentfulEvent} />
-        <EventsYouMayLike eventId={id} />
-      </PageWrapper>
-    )
-  }
+      </EventInfoCardWrapper>
+      <Container>
+        <Row>
+          <Column width={[1, 1, 0.6]}>
+            <TitleWrapper>
+              <Title>{name}</Title>
+              <EventTagList values={eventCategories} />
+            </TitleWrapper>
+            <Section>
+              <ReactMarkdown source={eventDescription} />
+            </Section>
+            {sponsorSection && (
+              <SponsorsSubsection title={sponsorSection.displayName}>
+                {renderSponsors(sponsorSection.sponsors, true)}
+              </SponsorsSubsection>
+            )}
+            {accessibilityDetails && (
+              <>
+                <AccessibilityHeading>Accessibility</AccessibilityHeading>
+                <Section>
+                  <ReactMarkdown
+                    source={accessibilityDetails.accessibilityDetails}
+                  />
+                </Section>
+              </>
+            )}
+            {performances && (
+              <Section>
+                <EventSchedule schedule={performances} />
+              </Section>
+            )}
+            <EventShareSection
+              name={name}
+              location={metaUrl}
+              date={startTime}
+            />
+          </Column>
+        </Row>
+      </Container>
+      <EventDirectionsSection data={contentfulEvent} />
+      <EventsYouMayLike eventId={id} />
+    </PageWrapper>
+  )
 }
 
 Event.propTypes = {
@@ -364,6 +356,8 @@ Event.propTypes = {
   location: PropTypes.object.isRequired,
   pageContext: PropTypes.object.isRequired,
 }
+
+export default Event
 
 export const eventPageQuery = graphql`
   query eventQuery($id: String!) {
