@@ -1,26 +1,27 @@
 import { css } from 'styled-components'
+import { sm, md, lg, xl } from './breakpoints'
 import theme from './theme'
 
 export const media = {
   // Min width media queries
   mobile: (...args) => css`
-    @media (min-width: ${theme.breakpoints[0]}px) {
+    @media (min-width: ${sm}px) {
       ${css(...args)};
     }
   `,
   tablet: (...args) => css`
-    @media (min-width: ${theme.breakpoints[1]}px) {
+    @media (min-width: ${md}px) {
       ${css(...args)};
     }
   `,
   desktop: (...args) => css`
-    @media (min-width: ${theme.breakpoints[2]}px) {
+    @media (min-width: ${lg}px) {
       ${css(...args)};
     }
   `,
 
   desktopHD: (...args) => css`
-    @media (min-width: ${theme.breakpoints[3]}px) {
+    @media (min-width: ${xl}px) {
       ${css(...args)};
     }
   `,
@@ -33,23 +34,23 @@ export const media = {
 
   // Max width media queries
   mobileMax: (...args) => css`
-    @media (max-width: ${theme.breakpoints[0] - 1}px) {
+    @media (max-width: ${sm - 1}px) {
       ${css(...args)};
     }
   `,
   tabletMax: (...args) => css`
-    @media (max-width: ${theme.breakpoints[1] - 1}px) {
+    @media (max-width: ${md - 1}px) {
       ${css(...args)};
     }
   `,
   desktopMax: (...args) => css`
-    @media (max-width: ${theme.breakpoints[2] - 1}px) {
+    @media (max-width: ${lg - 1}px) {
       ${css(...args)};
     }
   `,
 
   desktopHDMax: (...args) => css`
-    @media (max-width: ${theme.breakpoints[3] - 1}px) {
+    @media (max-width: ${xl - 1}px) {
       ${css(...args)};
     }
   `,
